Job_collector_orchestrator_service/requirements.txt
psycopg2-binary
pydantic
httpx
apify-client
